consolidate testing of costume
repeat the other make_* consolidation - pull both create & edit into a common file, remove tests that are really only a retest of bidding, and make a parent class with setUpTestData to centralize and avoid repetition of database activities.
